Basic canvas hook implementation for asynchronous callbacks to ancillary canvas components (e.g. plugins). Hooks currently available include after user creation, after enrollment, after course creation, after course activation, after quiz submission, and after a quiz snapshot is taken.
Callbacks are registered in the initializer of the component by
invoking the register_callback method of the CanvasHook instance
with parameters for the desired hook and the callback.

Delayed job handles the callback invocation. Instances of
CanvasHookJob, which are initialized with the callback
and the parameter for the callback, are inserted into the
default queue and processed.